fix: Unboxing is unnecessary as for Java 5 and later, and can safely be removed.

+++ src/main/java/gtPlusPlus/api/objects/minecraft/BTF_Inventory.java
@@ -124,7 +124,7 @@ public class BTF_Inventory implements ISidedInventory {
         int[] arg6 = new int[tList.size()];
 
         for (int i = 0; i < arg6.length; ++i) {
-            arg6[i] = ((Integer) tList.get(i)).intValue();
+            arg6[i] = (Integer) tList.get(i);
         }
 
         return arg6;
