+++ src/main/java/gtPlusPlus/plugin/villagers/entity/EntityBaseVillager.java
@@ -528,11 +528,11 @@ public class EntityBaseVillager extends EntityVillager {
         Tuple tuple = (Tuple) blacksmithSellingList.get(aItem);
         return tuple == null
                 ? 1
-                : (((Integer) tuple.getFirst()).intValue() >= ((Integer) tuple.getSecond()).intValue()
-                        ? ((Integer) tuple.getFirst()).intValue()
-                        : ((Integer) tuple.getFirst()).intValue()
-                                + aRand.nextInt(((Integer) tuple.getSecond()).intValue()
-                                        - ((Integer) tuple.getFirst()).intValue()));
+                : ((Integer) tuple.getFirst() >= (Integer) tuple.getSecond()
+                        ? (Integer) tuple.getFirst()
+                        : (Integer) tuple.getFirst()
+                                + aRand.nextInt((Integer) tuple.getSecond()
+                                        - (Integer) tuple.getFirst()));
     }
 
     public static void addPurchaseRecipe(MerchantRecipeList aTradeList, Item aItem, Random aRand, float aChance) {
@@ -560,10 +560,10 @@ public class EntityBaseVillager extends EntityVillager {
         Tuple tuple = (Tuple) villagersSellingList.get(aItem);
         return tuple == null
                 ? 1
-                : (((Integer) tuple.getFirst()).intValue() >= ((Integer) tuple.getSecond()).intValue()
-                        ? ((Integer) tuple.getFirst()).intValue()
-                        : ((Integer) tuple.getFirst()).intValue()
-                                + aRand.nextInt(((Integer) tuple.getSecond()).intValue()
-                                        - ((Integer) tuple.getFirst()).intValue()));
+                : ((Integer) tuple.getFirst() >= (Integer) tuple.getSecond()
+                        ? (Integer) tuple.getFirst()
+                        : (Integer) tuple.getFirst()
+                                + aRand.nextInt((Integer) tuple.getSecond()
+                                        - (Integer) tuple.getFirst()));
     }
 }
